fix: if-else, switch 대신 다형성 패턴 사용하기
